Load in Spacecenter/trackingstation, scan from here
Is this really all it takes to make scanning independent of both SCANsat
parts and the flight scene?

It seems to work with only a few caveats. The small map and instruments
windows are more dependent on SCANsat parts so they are turned off when
no SCANsat parts are there.