README.md FAQ: Explain FOV and "best" altitude
This is based upon my best understanding by reading `SCANcontroller.doScanPass()`. At first I thought that FOV was a virtual cone drawn underneath the instrument to the surface, but the main sensor loop seems to just walk lat/long from -FOV to +FOV and register the scans with `SCANUtil.registerPass()`, with there being a linear penalty for being below the "best" altitude.

As an aside, walking from -FOV to +FOV means it would appear the actual width of the swathes is *twice* the FOV; in other words, a 5° FOV results in swathes which are 10° (1/36th Kerbin) in size. However my understanding there might be off, as this is my first time looking at the SCANsat code.

In any case, this PR answers what the "best" altitude means and how FOV are calculated, which are frequently asked questions if you're me. :)